requirements.txt
opencv-python~=4.8.0.76
numpy~=1.24.3
gymnasium~=0.29.1
torch~=1.13.1+cu116
tensorboard~=2.13.0
pywin32~=306
qimage2ndarray~=1.10.0
scipy~=1.11.2
pyqt5~=5.15.9
pillow~=10.0.0
keyboard~=0.13.5
torchvision~=0.14.1+cu116
matplotlib~=3.7.2